Delete RegisterForm component and related files:
- Delete RegisterForm.tsx
- Delete styles.module.scss
- Rename Register directory to UpdateUserStateForm
- Rename registerFormStyle.ts to updateUserStateFormStyle.ts and update imports accordingly
- Update AppGate.tsx to import UpdateStateForm from the new location
- Update index.ts to export UpdateUserStateForm
