+++ frontend/src/components/UpdateUserStateForm/index.ts
@@ -1,2 +1,1 @@
-export * from './RegisterForm';
 export * from './UpdateUserStateForm';
